Decompile `rcat` `e_spike_room`
