+++ config/splat.pspeu.stcat.yaml
@@ -83,8 +83,7 @@ segments:
       - [0x28E20, c, cat/unk_3A164]
       - [0x296D8, c, cat/e_spike_room_light_switch]
       - [0x2A0C8, c, cat/unk_38AD0]
-      - [0x2A398, c, cat/e_spike_room]
-      - [0x2A9F0, c, cat/e_spikes]
+      - [0x2A398, c, cat/e_spikes]
       - [0x2B4D8, c, cat/e_elevator]
       - [0x2B980, .data, cat/e_stage_name]
       - [0x2C238, .data, cat/e_discus_lord]
@@ -124,8 +123,7 @@ segments:
       - [0x5F6A0, .data, cat/unk_3B4CC]
       - [0x5F6A8, .data, cat/unk_3A164]
       - [0x5F708, .data, cat/unk_38AD0]
-      - [0x5F718, .data, cat/e_spike_room]
-      - [0x5F728, .data, cat/e_spikes]
+      - [0x5F718, .data, cat/e_spikes]
       - [0x5F748, .data, cat/e_elevator]
       - [0x5F780, .rodata, cat/e_red_door]
       - [0x5F798, .rodata, cat/e_grave_keeper]
@@ -147,5 +145,5 @@ segments:
       - [0x5FCF0, bss, 5FCF0]
       - [0x608F0, .bss, cat/e_lock_camera]
       - [0x608F8, .bss, cat/unk_3A164]
-      - [0x60900, .bss, cat/e_spike_room]
+      - [0x60900, .bss, cat/e_spikes]
   - [0x60980]
