+++ config/splat.pspeu.strcat.yaml
@@ -68,8 +68,7 @@ segments:
       - [0x161D0, c, rcat/e_room_bg]
       - [0x162B0, c, rcat/e_lock_camera]
       - [0x16510, c, rcat_psp/unk_16510]
-      - [0x16778, c, rcat_psp/e_spike_room]
-      - [0x16BE0, c, rcat/e_spikes]
+      - [0x16778, c, rcat/e_spikes]
       - [0x176C8, c, rcat/e_cavern_background]
       - [0x18330, c, rcat/e_lava]
       - [0x19180, .data, rcat/e_red_door]
