+++ config/splat.us.stcat.yaml
@@ -52,8 +52,7 @@ segments:
       - [0x1340, .data, background_block_init]
       - [0x1358, .data, e_lock_camera]
       - [0x13D8, .data, e_breakable]
-      - [0x14F8, .data, e_spike_room]
-      - [0x14FC, .data, e_spikes]
+      - [0x14F8, .data, e_spikes]
       - [0x1518, .data, e_secrets]
       - [0x1620, .data, unk_3A164]
       - [0x1684, .data, e_coffin]
@@ -89,7 +88,7 @@ segments:
       - [0x2F528, data] # looks to be an unused sprite bank
       - [0x2FA64, .data, gen/us/sprites]
       - [0x366D8, .rodata, e_spike_room_light_switch]
-      - [0x366F8, .rodata, e_spike_room]
+      - [0x366F8, .rodata, e_spikes]
       - [0x36708, .rodata, unk_3872C]
       - [0x36718, .rodata, unk_38AD0]
       - [0x3673C, .rodata, e_elevator]
@@ -115,8 +114,7 @@ segments:
       - [0x36EC4, c, e_lock_camera]
       - [0x37080, c, e_breakable]
       - [0x3732C, c, e_spike_room_light_switch]
-      - [0x37A94, c, e_spike_room]
-      - [0x37F2C, c, e_spikes]
+      - [0x37A94, c, e_spikes]
       - [0x3872C, c, unk_3872C]
       - [0x38AD0, c, unk_38AD0]
       - [0x38D2C, c, e_elevator]
