+++ config/splat.us.strcat.yaml
@@ -80,8 +80,7 @@ segments:
       - [0x2FAEC, c, e_breakable]
       - [0x2FD98, c, e_lava]
       - [0x309C8, c, e_cavern_background]
-      - [0x31398, c, e_spike_room]
-      - [0x3171C, c, e_spikes]
+      - [0x31398, c, e_spikes]
       - [0x31F1C, c, unk_31F1C]
       - [0x32B98, c, e_secrets]
       - [0x33C34, c, e_bat]
