+++ config/symbols.us.strcat.txt
@@ -10,10 +10,7 @@ RCAT_BackgroundBlockInit = 0x801810EC;
 RCAT_RedDoorTiles = 0x80181114;
 RCAT_PrizeDrops = 0x801813C8;
 RCAT_rooms = 0x80182484;
-EntitySpikeRoomDarkness = 0x801B1398;
 func_us_801B87E8_from_cat = 0x801B2118;
 func_us_801BACF4_from_cat = 0x801B23D8;
-CreateEntityFromCurrentEntity = 0x801B65C8;
 DestroyEntity = 0x801B7368;
-AllocEntity = 0x801B7BD8;
 InitializeEntity = 0x801B8078;
