--- src/st/rcat/e_spike_room.c
@@ -1,4 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "rcat.h"
-
-INCLUDE_ASM("st/rcat/nonmatchings/e_spike_room", EntitySpikeRoomDarkness);
