--- src/st/rcat_psp/e_spike_room.c
@@ -1,4 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "../rcat/rcat.h"
-
-INCLUDE_ASM("st/rcat_psp/nonmatchings/rcat_psp/e_spike_room", EntitySpikeRoomDarkness);
